src/provider/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
python-dotenv==1.0.0
openai==1.58.0
httpx==0.28.1
